support multiple windows
there is still a global *widgets*, but now it is a hash table instead of
a list.  the ordered list of widgets for each window is in the 'widgets
property of the new class <window>.

widgets now have a property 'window to reference back to the window that
contains them.  this has to be set by 'widget-set-window!', and that step
is handled in 'make-window'.

<text-widget> now sets up its gc property in 'widget-set-window!' instead
of initialize-instance, because with the new design, the window is not
known at construction time.