implement -quit command line option
dbus:send does not work, in fact, but this is due to (depending on your
perspective) an error in the dbus egg, or an error in the documentation
for the dbus egg.  dbus:call does deliver the message successfully, but
because our quit handler does not return a value, dbus dumps out an error
message.  currently working to resolve the problem with dbus:send in the
dbus egg.