fix command line option "-quit" and corresponding dbus method
in the dbus egg, "dbus:register-method" corresponds to "dbus:call", while
"dbus:register-signal-handler" corresponds to "dbus:send".  'quit' properly
speaking, is a method, not a signal, so we use the former.  Another thing
we learned is that we can't use a continuation to quit the event loop
because dbus has to do some stuff after calling our method, namely send
a response.  So we have refactored the server's event loop a bit.