pass client-commands into server when we fork
this is a pretty horrible hack, as it requires a set of alternatives handlers
for all the client commands for when they are called by the server itself.
however, it does seem clear that the basic mechanism of passing the client
commands to the server will be necessary, as there is no apparent way to
get the timing right in running client commands as normal on the initial
invocation.