shell out to 'dbus-send' program instead of using dbus:discover-services
It seems to be impossible to fork a dbus program once you've done any
communication over dbus, including dbus:discover-services.  The child
process does not then get reliably assigned a unique bus name, causing
method calls from the parent (mowedline client) to timeout.  Experimentation
suggests that this is a problem in DBus itself, not just the dbus egg,
though not conclusively so.