sync client with spawned server so method calls don't happen too early
When the client forks to start the server, we need to ensure that the
server is ready to receive before sending any dbus method calls.  To
that end, we create a pipe, shared between parent and child process,
and have the child process write "ready\n" to the pipe when it is
initialized.