work on widget drawing
window-setup-widget-dimensions is now responsible for calculating the x
coordinate and width of widgets, instead of window-expose.  These values
are now stored in an xrectangle property of the widgets themselves as
well.

window-expose is now an ordinary procedure rather than a coops method to
allow us to use case-lambda.

<widget>.widget-set-window now initializes the gc for all subclasses of
<widget>, instead of only <text-widget>.

widget-width now returns #f if the widget is flexible.