parse-command-line: generalize
make parse-command-line agnostic about the details of the rest of the
application, namely, of specific command groups.  Additionally, the
structure it returns contains thunks for each of the commands, which can
be called directly, rather than the caller having to look up the name
of each command again.

define-command-group replaced by make-command-group.