new widget type <flags>
A flags widget is derived from text-widget, and serves the purpose of
displaying a set of flags.  The flags are pre-defined via its 'flags slot
and upon update, it parses the input to turn on, turn off, or clear and
replace the set of currently displayed flags.  The flags slot value is an
alist that looks like this:

      (("flagname" . displayspec)
       ...)

where flagname is a string token, as it will appear in the update text,
and displayspec is any valid text construct (a string or a colorized list
structure).

Input is parsed as follows.  If the input begins with "+", then the
following flagnames are turned on.  If the input begins with "-", then the
following flagnames are turned off.  Otherwise, the given flagnames
replace the current flag set.  Flagnames are given in a simple sequence,
separated by spaces.