event-loop: use srfi-18 'thread-sleep!' instead of 'usleep'
This is still far from perfect.  High CPU usage is dealt with, but at the cost
of laggy UI responsiveness.  What needs to be done, I learned, is to put each event
source in its own srfi-18 thread, and use 'thread-wait-for-i/o!' or the equivalent
for the library in question.