mowedline module
The mowedline server is now in a module 'mowedline', and the compilation
procedure changed so that the user config can import this module.  This
gives us the ability to set up the environment in which the config loads
by what the mowedline module exports.  More to follow in that way.

The build procedure is a little funny, but in summary, it uses the chicken
compiler to generate module registrations, and then builds those module
registrations into the mowedline binary.  This allows a 'load'ed script to
see the mowedline module at runtime.