Print an error if the config file can't be read
Set the default value of `user-config-file' to #f so that we can easily
distinguish between specified and unspecified config options.

Don't check for the existence of the given configuration file when
loading it, the command-line parsing should do that.

When parsing the command line see if we can read the given file and if
not print an error and exit with a failed status code.