x-eventloop: pass event to a method of the window it belongs to
This needs some further work, but the basic mechanism is as follows.  A
new method, window-handle-event does for a window what x-eventloop used
to do *to* windows.  By localizing event handling into the code of
windows, this should allow us to have more complex event handling in the
future.

The 'done' flag that used to be used to terminate the main event loop
has been replaced by setting up a continuation that exits it, and can be
called from anywhere.