x-eventloop rework
This patch reworks the way that X events are handled.  Instead of
x-eventloop itself handling events and operating on windows, the events
are passed to a method of <window> for local handling.  Provision is
made so that windows will be able to have local custom event handlers if
needed.

The quit mechanism is also affected.  Instead of a 'done' flag, there is
now a global quit continuation called 'quit-mowedline'.  This allows us
to quit the program from anywhere.

Buttons are also reworked a bit.  Previously buttons had a form that
when eval'd produced a thunk to call.  This was a mistake as there is no
reason to use eval here.  Now buttons have a handler which is a
procedure of one argument, widget.  Passing a widget reference to a
button when it is clicked allows the button to operate on its widget or
its window, expanding the capabilities of buttons in general.