widget-update(<map>): handle malformed input better
Map widgets expect to be able to read two symbols from their input.
Previously, if there were less than two symbols, ##eof would be read and
would appear in the widget's output.  Since the recent patch (242e7cd)
that applies the format function to the default text, this bug would
always occur on startup.  This patch deals with this type of malformed
input.