use xlib-utils event system
Replaces the global *windows* list with an xcontexts list, with our
window objects stored in xcontext-data.  The xlib-utils event system is
similar to mowedline's event system, but will allow us to listen to
events on non-mowedline windows (such as the root window) all with the
same event system, rather than having two event systems.  It's a work in
progress, so expect further patches.