create-mowedline-window renamed to window-create-xwindow, and use xcontext
Note, the foregoing patch and this patch are not, so far as I know,
actual valid uses of xcreatewindow, because they pass #f for the visual.
The program compiles and works though, so in subsequent patches we hope
to pass a correct value for visual, and then add support for window
transparency.  (The visual must have an alpha channel to do that.)

See also commit 6f20d686b89943d6078bc289f89723ebb7aec5b1, the original
commit where we switched from XCreateWindow to XCreateSimpleWindow due
to the mysterious error:

  X Error of failed request:  BadMatch (invalid parameter attributes)
        Major opcode of failed request:  1 (X_CreateWindow)
        Serial number of failed request:  10
        Current serial number in output stream:  13