version 0.2.9-dev
We're introducing a change to our versioning scheme that we hope to
adhere to with some diligence.  In future, after making a tagged
release, we will increment the version number again and add the suffix
"-dev".  This denotes a phase covering everything until the next tagged
release, at which we just drop the -dev suffix.  To clarify, X.Y.Z-dev
means older code than X.Y.Z.

This versioning scheme allows us to refer to unreleased code in
documentation and conversation without the ambiguity that there
currently is.

It should be possible to partially automate this process with a shell
script, and we will experiment with that when it is time for the next
tagged release.