Add proper dbus names to the service
It is convention to use a reversed domain to name the service, dbus
path and interface. It is also recommended against to use the path “/”
for dbus services because it makes versioning more difficult and makes
it hard for signal listeners to determine where the signal originated
from.

For introspection it seems necessary to respond to each path segment
separately. Testing with D-feet stopped working when the path was
introduced, but looking at other programs it seems to be necessary to
respond to each segment separately and name the next segment. For
example, the “/” points to the “/net” path:

    <node>
      <node name="net"/>
    </node>

And then the “/net” path points to the “/net/retroj” path.

    <node>
      <node name="retroj"/>
    </node>

Place the variables naming the path, service name and interface in a
separate file so both the client and the server can use it without
dublicating the information, in case we ever want to change it.  Keep
the introspection code in the ‘mowedline’ module because the client
doesn’t need it.