<window>: xinerama support
A mowedline window can be placed on a particular Xinerama screen by
providing the xinerama-screen property in current-xcontext when the window
is constructed.  The value of xinerama-screen is a xinerama-screen-info
record as obtained from xinerama-query-screens.

This patch does not provide a simple configuration api to do this, so that
will be added in a later patch.  With this patch, it can be accomplished
by parameterizing current-xcontext around the window constructor like
this:

(parameterize ((current-xcontext
                (make-xcontext
                 (current-xcontext)
                 xinerama-screen: (first
                                   (xinerama-query-screens
                                    (xcontext-display (current-xcontext)))))))
  (window ...))